pyproject.toml
[project]
name = "cv-hw2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.2.1",
    "torchvision==0.17.1",
    "torchaudio==2.2.1",
    "matplotlib==3.7.5",
    "numpy==1.24.0",
    "Pillow==10.4.0",
    "scikit-learn==1.3.2",
    "scipy==1.8.1",
    "tqdm==4.67.1",
    "gdown==5.2.0",
    "cython>=3.2.4",
    "setuptools>=82.0.1",
]

[tool.uv.extra-build-dependencies]
cyvlfeat = ["setuptools", "numpy", "cython"]
